refactor: move typed selection and famtypes
