+++ src/apps/ScopeChangeRequest/Components/Inputs/SearchableSingleSelect.tsx
@@ -8,7 +8,7 @@ import {
 } from 'react-select';
 import AsyncSelect from 'react-select/async';
 import { applyEdsComponents, applyEdsStyles, applyEDSTheme } from './SearchableDropdown/applyEds';
-import { TypedSelectOption } from '../../api/Search/searchType';
+import { TypedSelectOption } from '../../types/search/searchType';
 import { SearchableDropdownWrapper } from './SearchableDropdown/SearchableDropdownWrapper';
 import { useCancellationToken } from '../../../../hooks/cancellationToken/useCancellationToken';
