+++ src/apps/ScopeChangeRequest/Components/Workflow/Criteria/Components/AddContributor.tsx
@@ -7,7 +7,7 @@ import { Button, TextField } from '@equinor/eds-core-react';
 import { tokens } from '@equinor/eds-tokens';
 import { WorkflowIcon } from '../../Components/WorkflowIcon';
 import { useScopeChangeMutation } from '../../../../hooks/React-Query/useScopechangeMutation';
-import { TypedSelectOption } from '../../../../api/Search/searchType';
+import { TypedSelectOption } from '../../../../types/search/searchType';
 import { useScopeChangeContext } from '../../../../hooks/context/useScopeChangeContext';
 import { scopeChangeMutationKeys } from '../../../../keys/scopeChangeMutationKeys';
