+++ src/apps/ScopeChangeRequest/api/PCS/Batch/batchTag.ts
@@ -1,7 +1,7 @@
 import { httpClient } from '@equinor/lighthouse-portal-client';
 import { CommPkgProperty } from '../../../Components/Form/Inputs/ScopeChangeReferences/ScopeChangeReferences';
 import { Tag } from '../../../types/ProCoSys/Tag';
-import { TypedSelectOption } from '../../Search/searchType';
+import { TypedSelectOption } from '../../../types/search/searchType';
 
 export async function fetchBatchTags(
     tagNos: string[],
