+++ src/apps/ScopeChangeRequest/api/Search/PCS/searchNcr.ts
@@ -1,6 +1,6 @@
 import { isProduction } from '../../../../../Core/Client/Functions';
 import { HttpClient } from '../../../../../Core/httpClient/src';
-import { TypedSelectOption } from '../searchType';
+import { TypedSelectOption } from '../../../types/search/searchType';
 import { NCR } from '../../../types/ProCoSys/ncr';
 import { PCSStructure } from './searchStructure';
 
@@ -29,7 +29,8 @@ export const searchNCR = async (
     try {
         await client
             .fetch(
-                `api/Search?plantId=${encodeURIComponent(plantId)}&savedSearchId=${isProduction() ? searchIdProd : searchIdDev
+                `api/Search?plantId=${encodeURIComponent(plantId)}&savedSearchId=${
+                    isProduction() ? searchIdProd : searchIdDev
                 }&itemsPerPage=7&paging=true&sortColumns=false&api-version=4.1`,
                 requestOptions
             )
