+++ src/apps/ScopeChangeRequest/api/Search/PCS/searchPerson.ts
@@ -1,5 +1,5 @@
 import { HttpClient } from '@equinor/http-client';
-import { TypedSelectOption } from '../searchType';
+import { TypedSelectOption } from '../../../types/search/searchType';
 import { Person } from '../../../types/ProCoSys/person';
 
 export const searchPerson = async (
