+++ src/apps/ScopeChangeRequest/api/Search/PCS/searchQuery.ts
@@ -1,6 +1,6 @@
 import { HttpClient } from '@equinor/http-client';
 import { isProduction } from '../../../../../Core/Client/Functions';
-import { TypedSelectOption } from '../searchType';
+import { TypedSelectOption } from '../../../types/search/searchType';
 import { Query } from '../../../types/ProCoSys/query';
 import { PCSStructure } from './searchStructure';
 
@@ -30,7 +30,8 @@ export const searchQueryOrigin = async (
     try {
         await client
             .fetch(
-                `api/Search?plantId=${encodeURIComponent(plantId)}&savedSearchId=${isProduction() ? searchIdProd : searchIdDev
+                `api/Search?plantId=${encodeURIComponent(plantId)}&savedSearchId=${
+                    isProduction() ? searchIdProd : searchIdDev
                 }&itemsPerPage=7&paging=true&sortColumns=false&api-version=4.1`,
                 requestOptions
             )
