+++ src/apps/ScopeChangeRequest/api/Search/PCS/searchSWCR.ts
@@ -1,6 +1,6 @@
 import { HttpClient } from '@equinor/http-client';
 import { isProduction } from '../../../../../Core/Client/Functions';
-import { TypedSelectOption } from '../searchType';
+import { TypedSelectOption } from '../../../types/search/searchType';
 import { PCSStructure } from './searchStructure';
 import { SWCR } from '../../../types/ProCoSys/swcr';
 
@@ -30,7 +30,8 @@ export const searchSWCR = async (
     try {
         await client
             .fetch(
-                `api/Search?plantId=${encodeURIComponent(plantId)}&savedSearchId=${isProduction() ? searchIdProd : searchIdDev
+                `api/Search?plantId=${encodeURIComponent(plantId)}&savedSearchId=${
+                    isProduction() ? searchIdProd : searchIdDev
                 }&itemsPerPage=7&paging=true&sortColumns=false&api-version=4.1`,
                 requestOptions
             )
