+++ src/apps/ScopeChangeRequest/api/Search/STID/searchDocuments.ts
@@ -2,7 +2,7 @@ import { httpClient } from '../../../../../Core/Client/Functions/HttpClient';
 import { transformIsoDate } from '../../../Components/Workflow/Utils/dateFormatting';
 import { throwOnError } from '../../../functions/throwError';
 import { Document } from '../../../types/STID/document';
-import { TypedSelectOption } from '../searchType';
+import { TypedSelectOption } from '../../../types/search/searchType';
 
 export const searchDocuments = async (
     searchString: string,
@@ -29,8 +29,9 @@ export const searchDocuments = async (
             type: 'document',
             searchValue: x.docNo,
             object: x,
-            metadata: `Revision ${x.revNo} | Rev date ${x.revDate && transformIsoDate(x.revDate)
-                } | Reason for issue ${x.reasonForIssue}`,
+            metadata: `Revision ${x.revNo} | Rev date ${
+                x.revDate && transformIsoDate(x.revDate)
+            } | Reason for issue ${x.reasonForIssue}`,
         })
     );
 };
