+++ src/apps/ScopeChangeRequest/types/FAM/famTypes.ts
@@ -1,0 +1,1 @@
+export type FAMTypes = 'punch' | 'famtag' | 'htcable' | 'famtagno' | 'htcabletagno';
