+++ src/apps/ScopeChangeRequest/types/search/referenceType.ts
@@ -1,0 +1,5 @@
+import { FAMTypes } from '../FAM/famTypes';
+import { ProcoSysTypes } from '../ProCoSys/ProCoSysTypes';
+import { StidTypes } from '../STID/STIDTypes';
+
+export type ReferenceType = ProcoSysTypes | StidTypes | FAMTypes;
